src/gw-backend/requirements.txt
paho-mqtt==1.5.1
pyserial==3.5
configparser==5.0.1
